contrib: gpg-error: remove broken Android hacks
Proper handling of errors and interrupts of course requires that the
target only exists on succesful completion. .DELETE_ON_ERROR cannot be
used for directories. The contribs source README indeed has required to
keep $(MOVE) for as long as the current incarnation of contribs has
existed.

This hack is no longer necessary anyway as per the previous change.