aarch64: advertise assembler BTI support
As per the AArch64 ELF ABI, object files that support BTI need to
advertise it via a dedicated Linux ELF object file property in an ELF
note section. The linker can then check if all input object files
support the feature to advertise it at the shared object or executable
level.

In the case of BTI, this is necessary to enable the guarded page bit
in the memory mapping for BTI-capable .text sections.

This defines a macro to populate the note section and uses it.