vpx: switch mapping of option values to match vpx defines
This fixes the bad option default for this value. The default being
`VPX_DL_GOOD_QUALITY`, which is defined as `1000000` not `0` as seemed to be
the author's assumption, did not map to the specified range of 0-2 and thus
was getting clamped to 2, corresponding to 'best' quality rather than 'good'
quality.

It was requested in review to map the option values to match the vpx defines
rather than simply change the default value to `0`. This necessitates
removal of the range.