adjust: remove brightness threshold
No matter how you look at it, this did not belong in the image adjust
filter. This could belong in a separate module, or as a submodule of
the colour threshold filter -- if somebody actually ever needs it.

This function was never implemented by the hardware-accelerated adjust
filters, as it is indeed not a standard part of HSL and HSV adjustment
functionality, or really at all a part of it. So this has actually
become a functional inconsistency.