access: add a cache helper
Helper that can be used by any accesses, that will save up to 5 contexts
up to 5 seconds (both configurable with a define).

This helper create its own thread that will release the resources when
unused.

This helper also use the gcc destructor attribute to clean everything
when the library is unloaded.

This helper need a destructor, thus it can only be used with gcc and
clang.