lib: media_player: fix doxygen
Doxygen for libvlc_video_output_cfg_t and libvlc_video_render_cfg_t
were not correct and led to the documentation of one field being
assigned to the field after it.

In the future, the fields might need more precision regarding the
possible ranges or kind of values they can convey, so use a top
comment instead of a sibling one, which can more easily integrate
a brief and a longer explanation.