win32: sensors: refactor SUCCEEDED into FAILED
Early fail and leave the function or continue the loop when an error
happens. The code will less look like a tornado and will directly
signal what happens in case of error, without breaking the flow of the
current success path.

Indentation is not changed to ease review, and will be changed in
a subsequent commit.