qt/DirectCompositor: always insert external visual below all siblings
it seems the code assumed that null m_videoVisual means that video
visual is not in the visual tree but that assumption is wrong, this
caused the addition of acrylic visual to fail just after the video is
played and stopped. In that case the videoVisual will not be null but
also will not be in visual tree

Always insert acrylic visual below all siblings.
probably fixes #26279