opengl: implement vlc_gl_HasExtension in modules
Remove vlc_gl_HasExtension helper from the public API and implement it
as a private helper in modules. The reason for such helper was to help
ensuring the core stays compatible with Core OpenGL profile, which is
broken as soon as glGetString(GL_EXTENSIONS) is called instead of the
glGetStringi(GL_EXTENSION, i) variant.

Having the helper in the core led to issues against the MacOSX
implementation.

The extension store in gl_util.h prevents reloading the functions
across multiple calls from vlc_gl_HasExtension, and the helper can
easily be used in any OpenGL client which needs access to the extension
list.

Fix #26606 regression as a side effect

Co-Authored-by: Zhao Zhili <quinkblack@foxmail.com>