qt: fix keyboard navigation getting stuck in effects dialog sync tab
If you used the `Tab` key to navigate between controls within the
synchronization tab of the effects dialog, once you landed upon the refresh
button (the circular icon, top-right) navigation got stuck, refusing to
jump to a new control.

I tracked down the cause to be related to the fact that the next control in
the tab list will be the 'apply' button, but in a tab change event handler
we hide the apply button for the sync tab. This seems to confuse Qt, though
navigating in reverse (`Shift`+`Tab`) works fine. Flipping the focus policy
of the button depending upon whether or not it is hidden works as a fix.

Fixes #25923.