qt/messages: fix lack of focus border on button
The update/clear icon-based button here gives no indication that it has
the focus if you tab to it via keyboard navigation. Switching the button
type from `QPushButton` to `QToolButton` fixes this.

Note that the similar button in the sync tab of the effects dialog is of
type `QToolButton` and did not exhibit the problem.

(Non-functional change; no other change in appearance).