orient: factor composite transform plane functions
Rotations and anti-transposition are defined as the composition of
horizontal flip or transposition with one or two "free" vertical flips
(i.e. negated stride). The 3 functions, for 8-, 16- and 32-bit samples
were essentially identical.

This takes them out of the transforms macro and merge the function.
With this size-specific functions and function pointers remain only for
horizontal flip and transposition. This reduces the code size, and more
importantly simplifies writing SIMD-optimised transforms.