libplacebo: update for new tone mapping API
Libplacebo v4.192 got a very significant overhaul of the tone-mapping
subsystem and associated APIs. For consistency's sake, present users
with only a UI based on the new configuration style, and use
backwards-compatibility code to map the new enums back to the old API's
set of options.

To avoid having to write this boilerplate twice, provide a common helper
that both the libplacebo vout and the opengl vout can use.