cpu: probe by descending priorities
The callbacks are initially for default C implementation, and are
updated by the "worst" optimisation first, with the best last.
(We do not stop at the first supported optimisation as not all all
modules necessarily provide all optimised callbacks.)

This probes modules by descending score, so that we can assign highest
score to the best optimisation.

This makes no functonal differences, merely provide a more intuitive
module scoring.