libvlc: disable OSD by default
Most libvlc usage means displaying the video in their own UI with their own
style. We should not push our OSD style by default. It can still be added
back with an "osd" parameter when calling libvlc_new().

Keep the OSD on Desktop builds, although with the Qt integration it may be
better to do it in the UI layer as well.