qt: fix excessive base heights in simple prefs ui files
Viewing the ui files in Qt creator revealed excess empty space at the
bottom of some, a huge amount in the case of media-library. This ultimately
has no impact upon the final result, but I thought I'd fix it for neatness.

Note that these numbers were the product of simply dragging the canvas size
until it snapped into place.