v4l2: report user buffer not supported
User pointers and memory map modes are identified by the same device
capability flag (STREAMING). If user pointers are not supported (and
they typically are not for physical hardware), then REQBUFS returns
EINVAL.

If the error is EINVAL, printing the error message "Invalid arguments"
gives the wrong impression that VLC is misuing the V4L API. If the
error is something else, then it is a real error, so should be printed
as such.