qt: group simple prefs look-n-feel selection into sub-widget
Rather than have a grid covering the entire contents of the look-n-feel
groupbox, which creates unnecessary complexity, use a sub-widget to group
together and layout the radio controls and the line underneath, letting the
parent layout thus just be a simple vertical-box. Using a widget rather
than just a sub-layout is necessary if we are to be able to do things like
hide the new selection sub-widget when not applicable.

The difference is almost identical, the radio controls are just spaced a
few pixels further apart.

This puts us in a position to hide the selection control and preview image
if skinning is not available (the previous layout did not work for this).