v4l2: track in-flight buffers
This tracks which buffers have been dequeued and not yet requeued.
When the capture ends, only buffers which were still queued will be
freed. The pool is freed whence all buffers have been freed.

This also automatically requeues a buffer when it is freed as long as
the capture has not been terminated.