vout: interrupt wait on UI resize request
On resize, the UI thread waits for the end of any current rendering at
the old size before returning.

To limit its waiting, interrupt the wait between prepare() and display()
on the vout thread to complete the rendering as soon as possible. As a
consequence, the display() call might be performed early (before its
expected PTS).

Refs !324