access: cdda: remove CDDB support
The CDDB databse from freedb.videolan.org doesn't reply anymore (due to
MusicBrainz ?). freedb.org is dead and gnudb.org is empty.

We still have MusicBrainz which is better.

The one reason it may still work for some is that they run their own server.
And the CDs they have are not found in MusicBrainz.