opengl: interop_sw: implement as a module
This commit introduce the "opengl sw interop" to allow interop to be
implemented as modules for the desc->plane_count != 0 case.

No functional changes, but now the interop doesn't need to be built into
the clients that are using it and is separated into a different plugin.

It still doesn't allow to use software interop in --glinterop parameter.