smb2: destroy the context in case of error
This fixes a potential stack-buffer-overflow when destroying a context
from Close() if an operation was aborted. Indeed, the smb2_destroy()
function might trigger callbacks with private data that was allocated on
an old function stack. To fix this issue, always destroy the smb2
context immediately after an error (when the struct vlc_smb2_op is
valid).

This issue is currently hidden by the teardown mechanism (but still
possible), that always try to close gracefully in case of error.

(cherry picked from commit 924c951518f02dd05436d46323301702df516f4f)
Signed-off-by: Thomas Guillem <thomas@gllm.fr>