VLCVideoUIView: follow guidelines from WWDC21
When using timestamp vs targetTimestamp, one must take care to compare
the VSYNC against the previous VSYNC (so targetTimestamp vs previous
targetTimestamp) and not use timestamp + targetTimestamp.

Also get the current time through CoreAnimation time functions to build
the difference with now, allowing a bit of latency to be introduced in
timestamp.