opengl: explicitly set swizzle_per_tex values
As explained in commit 1e72174721e4af164b23e00578e097f029494260 (in
VLC4):

> In case the texture format is GL_LUMINANCE, swizzle_per_tex was set to
> {NULL, "xa"}. Using NULL instead of an explicit swizzle ("x") was a
> small optimization to assign the texel value directly [...]
>
> This was possible because the texture exposes the luminance value in
> its three components: (L, L, L, 1).

In fact, using the luminance value in the 3 components is not equivalent
(and wrong). Therefore, explicitly use the texture swizzle.

Cherry-picked from 1e72174721e4af164b23e00578e097f029494260

Note: only the assignment is cherry-picked, the associated refactor is
not applied to minimize code changes on VLC3.