aout: split aout and stream internally
This change is required for the gapless rework, to identify 2 different
streams, using the same aout APIS. The gapless stream, will be created,
if it has the same format as the previous one, and will wait in the
first Play to make a transition with the previous stream.

This is also a prerequisite for the aout/stream rework (in all aout
modules) planned for VLC 5.0.

Splitting the 2 API internally partition the decoder.c code: the decoder
should only handle the aout stream and not the aout devices (like sound,
mute, selection...).

For now, the aout reference and ownership is still kept in the decoder.c
because the aout and input_resource don't allow creating more than one
streams from one aout. This will be fixed in VLC 5.0.