contrib/gnutls: fix AArch64 build on appleOS
This introduces a patch to fix compilation on AArch64 on AppleOS, which
currently fails due to a compiler flag used by gnutls that is
unsupported by the latest version of clang (Apple clang version 13.1.6,
clang-1316.0.21.2) leading to a compiler crash if used.

This resolves the upstream tickets
https://gitlab.com/gnutls/gnutls/-/issues/1347 and
https://gitlab.com/gnutls/gnutls/-/issues/1317 for which the same patch
was already suggested.