opengl: expose GetTexFormatSize() in interop
This will allow to call this function from interop modules.

This function is currently called from opengl_interop_init_impl() (the
init function pointer), which will be removed (so this narrows the scope
of helpers code exposed to modules).

In the future, more work will be needed to also remove this function
pointer.