packetizer:av1: detect sequence header changes
When the sequence header changes we need to know it has changed so the output
values can be reset.

In particular the visible size was not reset since it was set initially by the
demuxer. The first sequence header received is still not used to overwrite the
visible size after this patch.

The extra data were not reset even though they contain the sequence header.

Because there changes were not detected the new output format is never
forwarded.

Fixes #26811