macosx: ignore DISPLAY_SIZE and handle resize internally
The current vout display module is resizing from its own view and tries
to notify the video view changes to the parent. But it can already
resize, so it doesn't need the core to resize the input picture, and
would already know the size it can render to.

It makes reporting the vout display size from the vout useless, and even
harmful since the display will re-enter itself on resize.