contrib: regex: Apply a patch to add missing function declarations
This code seems to be a stale copy of a regex library from glibc,
thus not sending the patch further upstream.

In some configurations, this library seems to choose not to include
standard C headers but declare the functions manually (without a
proper prototype though!). In this case, it already declared malloc
and realloc, but didn't declare abort and free in the same way.

Just add declarations of these functions in the same way - while
the most correct path forward would be to actually make it use
the proper system headers.

(cherry picked from commit 0e836b4eb2b05c8e130b73dcce8c53ac8397466f)