interop_vaapi: fix 420_10BPP format description
VLC_CODEC_VAAPI_420_10BPP is equivalent to VLC_FOURCC_P010 and has a
plane description that is like NV12, so it should not be different from
NV12. Only the bits per component, and so texture storage type is
changing for this format.

It was leading to incorrect or green color rendering with vaapi.
Regression from c344522ff264caa43c9dbc710484f01edbad04cb.

Fixes #26883