vdpau/mixer: remove useless code
If the input format is not some type of VDPAU YUV, then the output
format should not be VDPAU RGB, so this situation should never happen.

Nowadays the OpenGL or Vulkan display can handle YUV pictures nicely,
so there is anyway not much point left in using the VDPAU mixer without
the VDPAU decoder.