contrib: librist: patch meson.build for libcjson
Debian is using cJSON from github.com/DaveGamble/cJSON, as well as
archlinux and homebrew, which is shipping a libcjson.pc file. Only
ubuntu is using a different version not including the .pc file.

Homebrew is also shipping a libcjson.pc file, and without this, it would
generate a pc file like that:

    prefix=/foo/bar/vlc/contrib/x86_64-apple-darwin21
    libdir=${prefix}/lib
    includedir=${prefix}/include

    Name: librist
    Description: Reliable Internet Stream Transport (RIST)
    Version: 0.2.4
    Libs: -L${libdir} -lrist /usr/local/Cellar/cjson/1.7.15/lib/libcjson.1.7.15.dylib
    Cflags: -I${includedir} -I/usr/local/Cellar/cjson/1.7.15/include

Which won't be valid, especially for a standalone build with the
macosx/build.sh script.

It was leading to undefined references to cJSON symbols when it's
installed from homebrew.

Backport from upstream 809390b3b75a259a704079d0fb4d8f1b5f7fa956.

Refs #26860