vout/control: favour holders over wakers
Holders typically expect to be processed as soon as possible (e.g.
toggling pause) as response to user interaction. Waking up the control
queue on the other hand was only meant to avoid dead lock, and nowadays
only serves to avoid waiting for the control deadline.

Note that we do not really need to recheck the pending count in the
original wait loop. If we get there, we will wait for the time-out and
let holders pass anyway.