vlc_vout_display: add projection field and control
Even though the source provides a (source) projection, this new
parameters provides a way to specify how the projection should be
rendered on-screen. The current goal is not to provide projection
filtering (eg. changing from cubemap to equirectangular) but mostly to
be able to enforce a rectangular projection rendering and avoid the
mapping to 3D objects on the display.