vlc_fixup: fix usage of typeof from C++
Replace typeof by a decltype()-based equivalent on compilers not
supporting typeof.

Decltype is not equivalent to typeof, but we are mostly concerned about
removing references from the type, since it has no use in C code like
VLC list.

Note that this fixup is there to fix the usage of headers using typeof
in C++ code, not to provide "typeof" directly C++ to replace decltype.