contrib: upnp: disable webserver and device handling
VLC 3.0 uses libupnp exclusively as a client. LibUpnp internals are
quite opaque and it has been observed that a webserver thread is still
running on idle with open sockets even if the library is used as a
client only.

The recent bump to 1.14 introduces regressions in libupnp's webserver
socket initialization [1] on windows. Since the bug is still not
corrected upstream and we don't have any use of the webserver for 3.0
anyway, let's just disable it.

[1] https://github.com/pupnp/pupnp/issues/195

Fixes videolan/vlc#27001