configure: do not enable AltiVec in libvlccore
This was added ostensibly to fix compilation of the run-time AltiVec
test which was removed in bc146294cfeabde3a0c919883931c17e7c36630d.
Unfortunately, it causes the compiler to emit AltiVec instructions,
such that the executable crashes if AltiVec is not available.

Regression from e48d61955575634ef0324158cac88353e97d50b8.