configure.ac: remove x-test checks
x-test syntax (x$foo = xyes/xno) is used to circumvent the fact that
shell variables are substituted and not just defining an argument, so
that $foo doesn't expand to no argument at all.

But "" in posix shell syntax is considered as an argument by itself, so
using quotes around the variable removes this issue. Thus, x-test is
useless in the case where it has been removed in this patch.