macosx: disable --with-x on default configuration
When building Qt, we only add and link the X11 library and flags when
QT5_HAS_X11 is defined (in automake), which happens when we build
Qt5X11Extras.

Without this check, if autotools find the X11 library, then it might
fail to find the headers (because X_CFLAGS won't be used) or will at
least fail to link correct (because -lX11 will be missing from the
link command).

Since we don't use X11 on MacOSX, disable it by default in the configure
script like Qt, so that enabling Qt module locally works correctly.