caopengllayer: fix wrong fullscreen size
When transitioning to fullscreen, the codepath for non-live
resize was used and did not trigger rendering here.
As the size if now handled by the window however, this resulted
in a wrong size as it was never actually updated. To prevent that,
always render with the proper size from the layer regardless
if we are in live-resize or not.

Fixes #26962