codec: videotoolbox: handle kVTVideoDecoderReferenceMissingErr
The kVTVideoDecoderReferenceMissingErr error is not a critical one,
but without it being handled here, it would lead to aborting with
VideoToolbox and falling back to Software.

Reproduced this with a sample provided by zhilizhao(赵志立),
related to !1790.

I assume that before this error code was introduced in macOS 12/iOS 15,
VT did just return the NULL image buffer with a success status code,
leading to the issue seen in !1790.