video_output: flush static filters on picture drop
Filters like deinterlacers are using timestamps interpolation to
generate intermediate frames, which might not work reliably when the
timestamps are not from consecutive frames.

When dropping a decoded picture, we were not notifying those filters and
the next frame was arbitrarily further than the one the filter might
expect.

Ideally, we would notify which frames has been dropped to the filter, so
that it can differentiate a real discontinuity from a drop and can
handle its history correctly, but flushing is still better than doing
nothing and history is not really well exposed for now anyway.